[rename]: 세션 스토리지 파일 utils로 이동 (#236)
